+++ django_mapengine/layers.py
@@ -51,6 +51,11 @@ class ModelLayer:
     # pylint:disable=C0103
     id: str  # noqa: A003
     source: str
+    source_layer: Optional[str] = None
+
+    def get_source_layer(self) -> str:
+        """By default, layer ID is used as source layer."""
+        return self.source_layer if self.source_layer else self.id
 
 
 class StaticModelLayer(ModelLayer):
@@ -109,7 +114,7 @@ class StaticModelLayer(ModelLayer):
         yield MapLayer(
             id=self.id,
             source=self.source,
-            source_layer=self.id,
+            source_layer=self.get_source_layer(),
             minzoom=self.min_zoom(),
             maxzoom=self.max_zoom(),
             style=get_layer_style(self.id),
@@ -118,7 +123,7 @@ class StaticModelLayer(ModelLayer):
             yield MapLayer(
                 id=f"{self.id}_distilled",
                 source=f"{self.source}_distilled",
-                source_layer=self.id,
+                source_layer=self.get_source_layer(),
                 minzoom=self.min_zoom(distill=True),
                 maxzoom=self.max_zoom(distill=True),
                 style=get_layer_style(self.id),
@@ -204,17 +209,27 @@ def get_static_layers() -> Iterable[StaticModelLayer]:
     Return model layers for static-based MVTs.
 
     As multiple layers can have same source (via source_layer), API_MVTs is a dict where key is used as parent source.
+    Additionally, sources and source layers are reused if same model manager is used for multiple layers.
 
     Yields
     ------
     StaticModelLayer
         Static model layers to show models on map.
     """
-    for source, mvt_apis in settings.MAP_ENGINE_API_MVTS.items():
-        if source in settings.MAP_ENGINE_REGIONS:
-            continue
+    managers = {}
+    for original_source, mvt_apis in settings.MAP_ENGINE_API_MVTS.items():
         for mvt_api in mvt_apis:
-            yield StaticModelLayer(id=mvt_api.layer_id, source=source)
+            source_layer = mvt_api.layer_id
+            manager_reference = f"{mvt_api.model_name}.{mvt_api.manager_name}"
+            if manager_reference in managers:
+                # Add model managers only once and use source and source layer in multiple layers
+                source, source_layer = managers[manager_reference]
+            else:
+                managers[manager_reference] = (original_source, source_layer)
+                source = original_source
+            if original_source in settings.MAP_ENGINE_REGIONS:
+                continue
+            yield StaticModelLayer(id=mvt_api.layer_id, source=source, source_layer=source_layer)
 
 
 def get_cluster_layers() -> Iterable[ClusterModelLayer]: